refactor: Update Tutorial model schema and toJSON method
This commit refactors the Tutorial model schema in the `tututorialModel.js` file. The `trype` typo is corrected to `type` for both the `title` and `description` fields. Additionally, the `timestamps` option is set to `false` to disable automatic timestamp generation. The `toJSON` method is updated to remove the `_id` field from the serialized JSON object.