feat: Update department and personnel controllers
This commit updates the department.controller.js and personnel.controller.js files in the Personnel API.

In department.controller.js, the message for successfully creating a department is changed to "New Department successfully created" instead of "Department successfully created".

In personnel.controller.js, the message for successfully creating a personnel is updated based on whether the lead information is updated or not. If the lead information is updated, the message is "Old lead information is updated and new personnel successfully created". Otherwise, the message is "New personnel successfully created".

These changes aim to improve the clarity and accuracy of the messages in the API responses.