feat: Add TMDB_BASE_URL to environment variables
This commit adds the TMDB_BASE_URL to the environment variables in the movie app. The TMDB_BASE_URL is set to "https://api.themoviedb.org/3/movie/". This change allows the app to make API requests to the correct base URL when fetching movie data.

This enhancement improves the functionality of the app by ensuring that the correct base URL is used for API requests, enabling the retrieval of accurate movie data.